+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Shallow Route Challenge
 
 ## [1.0.0]
+
 ### Added
+
 - Initial commits.
+- Display required operation and route ids
+- Display required segment names
+
+### Changed
+
+- Render OperationPanel only for the route "operation"
+- Display "loading segment" between segment changes
+- Add loading operation for each operation change